[iOS] POST /api/match & /api/progress

+++ ios/baseball/baseball/Domain/Entity/Pitcher.swift
@@ -12,5 +12,5 @@ struct Pitcher: Codable {
     private (set) var numberOfPitching: Int
     private (set) var hit: Int
     private (set) var baseOnBalls: Int
-    private (set) var innings: Int
+    private (set) var innings: Float
 }
